+++ extensions/cornerstone-dicom-sr/src/commandsModule.js
@@ -128,13 +128,9 @@ const commandsModule = props => {
   const definitions = {
     downloadReport: {
       commandFn: actions.downloadReport,
-      storeContexts: [],
-      options: {},
     },
     storeMeasurements: {
       commandFn: actions.storeMeasurements,
-      storeContexts: [],
-      options: {},
     },
   };